% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{default_event_deck}
\alias{default_event_deck}
\title{Default Event Deck}
\format{
An object of class \code{character} of length 12.
}
\usage{
default_event_deck
}
\description{
The floristic relay game's default event deck
}
\keyword{datasets}
