% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{default_gameboard}
\alias{default_gameboard}
\title{Default gameboard}
\format{
An object of class \code{FloristicRelayGame::floristic_gameboard} (inherits from \code{S7_object}) of length 1.
}
\usage{
default_gameboard
}
\description{
The default floristic relay \code{\link[=floristic_gameboard]{gameboard}}
}
\keyword{datasets}
