% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{default_interaction_deck}
\alias{default_interaction_deck}
\title{Default Interaction Deck}
\format{
An object of class \code{list} of length 6.
}
\usage{
default_interaction_deck
}
\description{
The floristic relay game's default interaction deck
}
\keyword{datasets}
