% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{default_players}
\alias{default_players}
\title{Default Players}
\format{
An object of class \code{list} of length 6.
}
\usage{
default_players
}
\description{
A list of 6 players for the floristic relay game using each of the 6 pre-defined plants
}
\keyword{datasets}
