% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_response.R
\name{event_response}
\alias{event_response}
\title{Define a floristic relay plant's event response}
\usage{
event_response(event = character(0), response = integer(0))
}
\arguments{
\item{event}{A character string representing the name of the event. This should EXACTLY MATCH the name of an event in the event deck}

\item{response}{An integer representing how many spaces the plant moves in response to the event (can be positive or negative).}
}
\value{
An event_response object.
}
\description{
When an event card is drawn, a \code{\link[=floristic_plant]{plant}} responds to it by moving on the board. Every \code{\link[=floristic_plant]{plant}}
should have an event response defined for every event in the event deck. These are exclusively used
when defining \code{\link[=floristic_plant]{plant}}, so see that page for examples.
}
\seealso{
\link{floristic_plant}
}
