% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{facilitation_with_nutrients}
\alias{facilitation_with_nutrients}
\title{Facilitation with nutrients}
\format{
An object of class \code{FloristicRelayGame::floristic_interaction} (inherits from \code{S7_object}) of length 1.
}
\usage{
facilitation_with_nutrients
}
\description{
One of the default floristic relay \code{\link[=floristic_interaction]{interactions}}
}
\keyword{datasets}
