% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floristic_gameboard.R
\name{floristic_gameboard}
\alias{floristic_gameboard}
\title{Define floristic relay game board.}
\usage{
floristic_gameboard(spaces = integer(0))
}
\arguments{
\item{spaces}{A vector of integer numbers representing the gameboard.}
}
\value{
A floristic_gameboard object.
}
\description{
The board is linear and composed of spaces. These spaces have
integer values: most spaces are 0 (no effect), positive (e.g. 1 sends the player that
lands there forward one space), and negative (e.g. -2 sends the player that lands there
backward 2 spaces). The final space is the goal space and must be 0. The start space
is not included in this definition and is always implicitly 0.
}
\examples{
\dontrun{
default_gameboard <- floristic_gameboard(spaces = as.integer(c(0, 0, 0, 2, 0,
                                                               0, 0, -1, 0, 0,
                                                               0, 0, -1, 0, 0,
                                                               -1, 0, 0, 0, 2,
                                                               0, 0, -1, 0, 0)))
}
}
