% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{grickle_grass}
\alias{grickle_grass}
\title{Grickle grass}
\format{
An object of class \code{FloristicRelayGame::floristic_plant} (inherits from \code{S7_object}) of length 1.
}
\usage{
grickle_grass
}
\description{
One of the default floristic relay \code{\link[=floristic_plant]{plants}}
}
\keyword{datasets}
