% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{lorax_tree}
\alias{lorax_tree}
\title{Lorax tree}
\format{
An object of class \code{FloristicRelayGame::floristic_plant} (inherits from \code{S7_object}) of length 1.
}
\usage{
lorax_tree
}
\description{
One of the default floristic relay \code{\link[=floristic_plant]{plants}}
}
\keyword{datasets}
