% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{mimsy_bush}
\alias{mimsy_bush}
\title{Mimsy bush}
\format{
An object of class \code{FloristicRelayGame::floristic_plant} (inherits from \code{S7_object}) of length 1.
}
\usage{
mimsy_bush
}
\description{
One of the default floristic relay \code{\link[=floristic_plant]{plants}}
}
\keyword{datasets}
