% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{momerath_herb}
\alias{momerath_herb}
\title{Momerath herb}
\format{
An object of class \code{FloristicRelayGame::floristic_plant} (inherits from \code{S7_object}) of length 1.
}
\usage{
momerath_herb
}
\description{
One of the default floristic relay \code{\link[=floristic_plant]{plants}}
}
\keyword{datasets}
