% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_floristic_game.R
\name{simulate_floristic_game}
\alias{simulate_floristic_game}
\title{Simulate the floristic relay game}
\usage{
simulate_floristic_game(
  gameboard,
  event_deck,
  interactions_deck,
  player_order,
  turn_by_turn = FALSE
)
}
\arguments{
\item{gameboard}{A \code{\link[=floristic_gameboard]{floristic_gameboard}} object}

\item{event_deck}{A vector of character strings representing disturbance events. These MUST EXACTLY MATCH the "event" field of
each \code{\link[=floristic_plant]{plant's}} \code{\link[=event_response]{event responses}}. For example, \link{momerath_herb}
has an event response defined for the event "Wildfire", so it is valid to include "Wildfire" in an event_deck. See example below
as a demonstration of building a full deck. If you add an event that does not exactly match the responses defined for the plants
in the game, the function will fail. Once the deck is "depleted", it is cycled through in the same order.}

\item{interactions_deck}{A vector of \code{\link[=floristic_interaction]{floristic interactions}} that are drawn when players
are on the same square at the end of a turn. Players move according to the instructions on that card. Once the deck is
"depleted", it is cycled through in the same order.}

\item{player_order}{A list of \code{\link[=floristic_player]{players}}. Players play every turn in the same order as
this list. See example below for creating a game with 6 players.}

\item{turn_by_turn}{Logical. Print messages describing the game being simulated turn-by-turn (useful for debugging).
FALSE by default.}
}
\value{
A dataframe of the final rankings. It has the columns "Plant_Name", "Late_or_Early", "Final_Position", and "Rank".
\describe{
  \item{`Plant_Name`}{The name of the a \link{floristic_plant}.}
  \item{`Late or Early`}{The plant's "type"; either early-succession or late-succession; denoted "Early" and "Late", respectively.}
  \item{`Final_Position`}{The position of the plant on the \link{floristic_gameboard} at the end of the game. At least one will be >= length(gameboard@spaces).}
  \item{`Rank`}{The plant's rank. Rank 1 is the winner and they are enumerated from 1 to the number of total players in the game. Note that this order is determined by
     simply calling \code{\link[dplyr]{arrange}} on the Final_Position, so ties are not explicitly handled.}
}
}
\description{
A function to simulate a game of floristic relay. This function requires a \code{\link[=floristic_gameboard]{gameboard}},
 a vector of events (see param definition below), a vector of \code{\link[=floristic_interaction]{interactions}},
 and a list of \code{\link[=floristic_player]{players}}. The function returns a dataframe recording the results of
 the simulated game. If 'turn-by-turn' is set to TRUE, it prints messages detailing the game turn-by-turn (used for debug).
}
\examples{
\dontrun{
# Create and randomize events deck
# (events are drawn in this order; repeats after cycling through the deck)
# Can add in events of specific type here;
# e.g. sample(c(event_set, rep("No Disturbance", times = 6)))
event_deck <- c(rep("No Disturbance", times = 6),
                rep("Landslide", times = 2),
                rep("Animals Grazing", times = 2),
                rep("Wildfire", times = 2))
event_deck <- sample(event_deck)

# Create and randomize interactions deck (using pre-defined interactions)
# (interactions are drawn in this order; repeats after cycling through the deck)
interactions_deck <- c(competition_for_water,
                       competition_for_light,
                       facilitation_with_nutrients,
                       facilitation_with_shade,
                       tolerance,
                       tolerance)
interactions_deck <- sample(interactions_deck)

# Create and randomize player list (players play their turn in this order)
# (uses pre-defined floristic_plants)
player_order <- list(
    floristic_player(player_character = momerath_herb,
                     position = 0L),
    floristic_player(player_character = lorax_tree,
                     position = 0L),
    floristic_player(player_character = grickle_grass,
                     position = 0L),
    floristic_player(player_character = truffula_tree,
                     position = 0L),
    floristic_player(player_character = mimsy_bush,
                     position = 0L),
    floristic_player(player_character = borogrove_grass,
                     position = 0L)
)
player_order <- sample(player_order)

# simulate game
# (Uses the default_gameboard)
game_result <- simulate_floristic_game(default_gameboard,
                                       event_deck,
                                       interactions_deck,
                                       player_order)
game_result
}
}
\seealso{
\link{floristic_player}, \link{floristic_gameboard}, \link{floristic_interaction}
}
