% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{tolerance}
\alias{tolerance}
\title{Tolerance}
\format{
An object of class \code{FloristicRelayGame::floristic_interaction} (inherits from \code{S7_object}) of length 1.
}
\usage{
tolerance
}
\description{
One of the default floristic relay \code{\link[=floristic_interaction]{interactions}}
}
\keyword{datasets}
